fixed keyboard

+++ src/enigma_machine.cpp
@@ -4,7 +4,7 @@
 Rotor::Rotor(std::string modelName, std::string symbols)
     : modelName_(modelName) {
   if (symbols.length() >= symbols_.size()) {
-    for (int i = 0; i < symbols_.size(); ++i) {
+    for (size_t i = 0; i < symbols_.size(); ++i) {
       symbols_[i] = symbols[i];
     }
   }
